+++ ascon-xof128/CHANGELOG.md
@@ -8,8 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Internal implementation by removing unnecessary buffering ([#849])
 - Serialization format used by `SerializableState` implementations ([#849])
+- Replaced implementation of `CustomizedInit` with `TryCustomizedInit` ([#857])
 
 [#849]: https://github.com/RustCrypto/hashes/pull/849
+[#857]: https://github.com/RustCrypto/hashes/pull/857
 
 ## 0.1.0 (2026-04-24)
 - Initial release ([#841])